c++ API for dump replacement
