+++ src/runtime_src/core/common/api/xrt_device.cpp
@@ -27,6 +27,9 @@
 #include "core/common/device.h"
 #include "core/common/message.h"
 #include "core/common/sensor.h"
+#include "core/common/info_memory.h"
+#include "core/common/info_platform.h"
+#include "core/common/info_host.h"
 #include "core/common/query_requests.h"
 
 #include "xclbin_int.h" // Non public xclbin APIs
@@ -284,14 +287,22 @@ get_info(info::device param) const
       (handle.get(), [](const auto& val) { return bool(val); });
   case info::device::offline :
     return query::raw<info::device::offline, xrt_core::query::is_offline>(handle.get());
-  case info::device::power_rails :            // std::string
-    return query::json_str(xrt_core::sensor::read_power_rails(handle.get()));
-  case info::device::thermals :               // std::string
+  case info::device::electrical :            // std::string
+    return query::json_str(xrt_core::sensor::read_electrical(handle.get()));
+  case info::device::thermal :               // std::string
     return query::json_str(xrt_core::sensor::read_thermals(handle.get()));
-  case info::device::power_consumption :      // std::string
-    return query::json_str(xrt_core::sensor::read_power_consumption(handle.get()));
-  case info::device::fans :                   // std::string
-    return query::json_str(xrt_core::sensor::read_fans(handle.get()));
+  case info::device::mechanical :                   // std::string
+    return query::json_str(xrt_core::sensor::read_mechanical(handle.get()));
+  case info::device::memory :                   // std::string
+    return query::json_str(xrt_core::memory::memory_topology(handle.get()));
+  case info::device::platform :                   // std::string
+    return query::json_str(xrt_core::platform::platform_info(handle.get()));
+    case info::device::pcie_info :                   // std::string
+    return query::json_str(xrt_core::platform::pcie_info(handle.get()));
+  case info::device::host :                   // std::string
+    boost::property_tree::ptree pt;
+    xrt_core::get_xrt_build_info(pt);
+    return query::json_str(pt);
   }
 
   throw std::runtime_error("internal error: unreachable");
