+++ src/runtime_src/core/tools/common/ReportPlatforms.cpp
@@ -20,77 +20,9 @@
 
 // Local - Include Files
 #include "ReportPlatforms.h"
-#include "XBUtilities.h"
-#include "core/common/query_requests.h"
-#include "core/common/device.h"
-namespace qr = xrt_core::query;
 
-static std::map<int, std::string> p2p_config_map = {
-  { 0, "disabled" },
-  { 1, "enabled" },
-  { 2, "error" },
-  { 3, "reboot" },
-  { 4, "not supported" },
-};
-
-/**
- * New flow for exposing mac addresses
- * qr::mac_contiguous_num is the total number of mac addresses
- * avaliable contiguously starting from qr::mac_addr_first
- * 
- * Old flow: Query the four sysfs nodes we have and validate them
- * before adding them to the property tree
- */
-static boost::property_tree::ptree
-mac_addresses(const xrt_core::device * dev)
-{
-  boost::property_tree::ptree ptree;
-  uint64_t mac_contiguous_num = 0;
-  std::string mac_addr_first;
-  try {
-    mac_contiguous_num = xrt_core::device_query<qr::mac_contiguous_num>(dev);
-    mac_addr_first = xrt_core::device_query<qr::mac_addr_first>(dev);
-  }
-  catch (const xrt_core::query::no_such_key&) {
-    // Ignoring if not available: Edge Case 
-  }
-
-  //new flow
-  if (mac_contiguous_num && !mac_addr_first.empty()) {
-    std::string mac_prefix = mac_addr_first.substr(0, mac_addr_first.find_last_of(":"));
-    std::string mac_base = mac_addr_first.substr(mac_addr_first.find_last_of(":") + 1);
-    std::stringstream ss;
-    uint32_t mac_base_val = 0;
-    ss << std::hex << mac_base;
-    ss >> mac_base_val;
-
-    for (uint32_t i = 0; i < (uint32_t)mac_contiguous_num; i++) {
-      boost::property_tree::ptree addr;
-      auto base = boost::format("%02X") % (mac_base_val + i);
-      addr.add("address", mac_prefix + ":" + base.str());
-      ptree.push_back(std::make_pair("", addr));
-    } 
-  }
-  else { //old flow
-    std::vector<std::string> mac_addr;
-    try {	  
-      mac_addr = xrt_core::device_query<qr::mac_addr_list>(dev);
-    }
-    catch (const xrt_core::query::no_such_key&) {
-      // Ignoring if not available: Edge Case 
-    }
-    for (const auto& a : mac_addr) {
-      boost::property_tree::ptree addr;
-      if (!a.empty() && a.compare("FF:FF:FF:FF:FF:FF") != 0) {
-        addr.add("address", a);
-        ptree.push_back(std::make_pair("", addr));
-      }
-    }
-  }
-
-  return ptree;
-}
-
+// 3rd Party Library - Include Files
+#include <boost/property_tree/json_parser.hpp>
 void
 ReportPlatforms::getPropertyTreeInternal( const xrt_core::device * dev, 
                                               boost::property_tree::ptree &pt) const
@@ -104,97 +36,14 @@ void
 ReportPlatforms::getPropertyTree20202( const xrt_core::device * dev, 
                                            boost::property_tree::ptree &pt) const
 {
-  boost::property_tree::ptree ptree;
+  xrt::device device(dev->get_device_id());
   boost::property_tree::ptree pt_platform;
-  
-  boost::property_tree::ptree static_region;
-  static_region.add("vbnv", xrt_core::device_query<qr::rom_vbnv>(dev));
-  try {
-    static_region.add("jtag_idcode", qr::idcode::to_string(xrt_core::device_query<qr::idcode>(dev)));
-  }
-  catch (const xrt_core::query::no_such_key&) {
-    // Ignoring if not available: Edge Case
-    static_region.add("jtag_idcode", "N/A");
-  }
-
-  try {
-    static_region.add("fpga_name", xrt_core::device_query<qr::rom_fpga_name>(dev));
-  }
-  catch (const xrt_core::query::no_such_key&) {
-    // Ignoring if not available: Edge Case 
-    static_region.add("fpga_name", "N/A");
-  }
-  
-  pt_platform.put_child("static_region", static_region);
-
-  boost::property_tree::ptree bd_info;
-  auto ddr_size_bytes = [](uint64_t size_gb, uint64_t count) {
-    auto bytes = size_gb * 1024 * 1024 * 1024;
-    return bytes * count;
-  };
-  bd_info.add("ddr_size_bytes", ddr_size_bytes(xrt_core::device_query<qr::rom_ddr_bank_size_gb>(dev), xrt_core::device_query<qr::rom_ddr_bank_count_max>(dev)));
-  bd_info.add("ddr_count", xrt_core::device_query<qr::rom_ddr_bank_count_max>(dev));
-  pt_platform.put_child("off_chip_board_info", bd_info);
-
-  boost::property_tree::ptree status;
-  try {
-    status.add("mig_calibrated", xrt_core::device_query<qr::status_mig_calibrated>(dev));
-  }
-  catch (const xrt_core::query::no_such_key&) {
-    // Ignoring if not available: Edge Case 
-    status.add("mig_calibrated", "N/A");
-  }
-
-  std::string msg;
-  auto value = XBUtilities::check_p2p_config(dev, msg);
-  status.add("p2p_status", p2p_config_map[value]);
-  pt_platform.put_child("status", status);
-
-  boost::property_tree::ptree controller;
-  boost::property_tree::ptree sc;
-  boost::property_tree::ptree cmc;
-  try {
-    sc.add("version", xrt_core::device_query<qr::xmc_sc_version>(dev));
-    sc.add("expected_version", xrt_core::device_query<qr::expected_sc_version>(dev));
-    cmc.add("version", xrt_core::device_query<qr::xmc_version>(dev));
-    cmc.add("serial_number", xrt_core::device_query<qr::xmc_serial_num>(dev));
-    cmc.add("oem_id", XBUtilities::parse_oem_id(xrt_core::device_query<qr::oem_id>(dev)));
-  }
-  catch (const xrt_core::query::no_such_key&) {
-    // Ignoring if not available: Edge Case 
-  }
-  controller.put_child("satellite_controller", sc);
-  controller.put_child("card_mgmt_controller", cmc);
-  pt_platform.put_child("controller", controller);
-
-  std::vector<char> raw; 
-  try { 
-    raw = xrt_core::device_query<qr::clock_freq_topology_raw>(dev);
-  }
-  catch (const xrt_core::query::no_such_key&) {
-    // Ignoring if not available: Edge Case 
-  }
-  if(!raw.empty()) {
-    boost::property_tree::ptree pt_clocks;
-    auto clock_topology = reinterpret_cast<const clock_freq_topology*>(raw.data());
-    for(int i = 0; i < clock_topology->m_count; i++) {
-      boost::property_tree::ptree clock;
-      clock.add("id", clock_topology->m_clock_freq[i].m_name);
-      clock.add("description", XBUtilities::parse_clock_id(clock_topology->m_clock_freq[i].m_name));
-      clock.add("freq_mhz", clock_topology->m_clock_freq[i].m_freq_Mhz);
-      pt_clocks.push_back(std::make_pair("", clock));
-    }
-    pt_platform.put_child("clocks", pt_clocks);
-  }
-
-  auto macs = mac_addresses(dev);
-  if(!macs.empty())
-    pt_platform.put_child("macs", macs);
-    
-  ptree.push_back(std::make_pair("", pt_platform));
+  std::stringstream ss;
+  ss << device.get_info<xrt::info::device::platform>();
+  boost::property_tree::read_json(ss, pt_platform);
  
   // There can only be 1 root node
-  pt.add_child("platforms", ptree);
+  pt.add_child("platforms", pt_platform);
 }
 
 void
