+++ src/runtime_src/core/tools/common/ReportThermal.cpp
@@ -35,7 +35,7 @@ ReportThermal::getPropertyTree20202( const xrt_core::device * _pDevice,
 {
   xrt::device device(_pDevice->get_device_id());
   std::stringstream ss;
-  ss << device.get_info<xrt::info::device::thermals>();
+  ss << device.get_info<xrt::info::device::thermal>();
   boost::property_tree::read_json(ss, _pt);
 }
